// docker/back-end/src/types/seller.ts
export type SellerType = {
  sellerId: number;
  fullName: string;
  password: string;
  birthday: string;
  cpf: string;
  nacionality: string;
  phoneNumber: string;
  email: string;
  gender: string;
  jobOccupation: string;
  bank: string;
  commission: number;
  salesQuantity: number;
  addressId: number;
};